requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
fastapi==0.115.12
httpx==0.28.1
huggingface-hub==0.30.2
jiter==0.9.0
pydantic==2.11.3
PyDispatcher==2.0.7
python-dotenv==1.1.0
requests==2.32.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
